broadcast: Filter equal messages and serialize messages
Using go functions within the broadcast itself is a rather bad
idea s the state of the broadcast might change while a go func
is being scheduled. As there is no guarantee on the order when
a go func is being scheduled a client may observe a new state
before an old state which would result in an inconsistent view
of the current ServerList.

The new channel filter ensures that no two same lists are
broadcasted. This happens quite often as Kubernetes may push
changes that do not modify the effective server list. In that
case all clients would reveive a state they already know. This
isn't necessary and pollutes only the logs.