lb: Perform round-robin load balancing in server list
The client will query the servers in a round-robin fashion with the
order of the server list it receives. We order the server list
differently for each client so that one-off calls do not hit always
the same server.

In order to increment the offset for each client we use an atomic
operation the increments a counter and resets its value when it wraps
around. The respective code is in the internal/atomic package.